Remove standalone node
