+++ .github/workflows/docker-hub.yml
@@ -1,4 +1,4 @@
-name: Docker Hub
+name: Publish features parachain to Docker Hub
 
 on:
   workflow_dispatch:
@@ -14,17 +14,33 @@ jobs:
       - name: Checkout repository
         uses: actions/checkout@v2
 
+      - name: Docker meta
+        id: meta
+        uses: docker/metadata-action@v3
+        with:
+          flavor: |
+            latest=true
+          images: |
+            zeitgeistpm/zeitgeist-node-parachain
+          tags: |
+            type=sha
+
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@v1
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v1
+
       - name: Login to DockerHub
         uses: docker/login-action@v1 
         with:
           username: ${{ secrets.DOCKERHUB_USERNAME }}
           password: ${{ secrets.DOCKERHUB_PASSWORD }}
-      
-      - name: Build and publish
-        run: |
-          # If it is a tag:
-          if [ -z "${GITHUB_REF##refs/tags/*}" ] ; then
-            ./scripts/docker-hub-publish.sh ${GITHUB_REF#refs/tags/}
-          else
-            ./scripts/docker-hub-publish.sh ${GITHUB_SHA}
-          fi
+
+      - name: Build and push
+        uses: docker/build-push-action@v2
+        with:
+          push: true
+          tags: ${{ steps.meta.outputs.tags }}
+          labels: ${{ steps.meta.outputs.labels }}
+
