+++ node/res/battery-station.json
