+++ node/src/chain_spec/dev.rs
@@ -1,7 +1,5 @@
-#[cfg(feature = "parachain")]
-use crate::chain_spec::get_from_seed;
 use crate::chain_spec::{
-    generic_genesis, get_account_id_from_seed, token_properties, zeitgeist_wasm,
+    generic_genesis, get_account_id_from_seed, get_from_seed, token_properties, zeitgeist_wasm,
     AdditionalChainSpec, ChainSpec,
 };
 use sc_service::ChainType;
@@ -10,9 +8,7 @@ use zeitgeist_primitives::types::Balance;
 
 const INITIAL_BALANCE: Balance = Balance::MAX >> 4;
 
-pub fn dev_config(
-    #[cfg(feature = "parachain")] parachain_id: cumulus_primitives_core::ParaId,
-) -> Result<ChainSpec, String> {
+pub fn dev_config(parachain_id: cumulus_primitives_core::ParaId) -> Result<ChainSpec, String> {
     let zeitgeist_wasm = zeitgeist_wasm()?;
     Ok(ChainSpec::from_genesis(
         "Development",
@@ -20,7 +16,6 @@ pub fn dev_config(
         ChainType::Local,
         move || {
             generic_genesis(
-                #[cfg(feature = "parachain")]
                 AdditionalChainSpec {
                     candidates: vec![(
                         get_account_id_from_seed::<sr25519::Public>("Alice"),
@@ -31,10 +26,6 @@ pub fn dev_config(
                     inflation_info: crate::chain_spec::DEFAULT_COLLATOR_INFLATION_INFO,
                     nominations: vec![],
                     parachain_id,
-                },
-                #[cfg(not(feature = "parachain"))]
-                AdditionalChainSpec {
-                    initial_authorities: vec![crate::chain_spec::authority_keys_from_seed("Alice")],
                 },
                 vec![
                     get_account_id_from_seed::<sr25519::Public>("Alice"),
@@ -59,12 +50,9 @@ pub fn dev_config(
         None,
         None,
         Some(token_properties()),
-        #[cfg(feature = "parachain")]
         crate::chain_spec::Extensions {
             relay_chain: "rococo-dev".into(),
             parachain_id: parachain_id.into(),
         },
-        #[cfg(not(feature = "parachain"))]
-        Default::default(),
     ))
 }
