+++ scripts/tests/all-sequencial.sh
@@ -5,7 +5,6 @@
 # IMPORTANT: CI verifies most of the following scripts in parallel
 
 . "$(dirname "$0")/misc.sh" --source-only
+. "$(dirname "$0")/runtime-and-node.sh" --source-only
 . "$(dirname "$0")/clippy.sh" --source-only
-. "$(dirname "$0")/parachain.sh" --source-only
-. "$(dirname "$0")/standalone.sh" --source-only
 . "$(dirname "$0")/fuzz.sh" --source-only
